requirements.txt
streamlit
langchain
langchain-openai
langchain-community
python-dotenv
pypdf
chromadb
tiktoken
